globalizando el cambio al español del tipo
